requirements.txt
asgiref==3.6.0
Django==4.1.9
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
